add test for importing pdk with data source and reorging a few tests
